[feat] 이미 북마크 중인지 확인하는 메소드 추가
+ bookmarkDTO postingId -> postId로 변경

+++ src/main/java/org/omoknoone/onionhotsayyo/bookmark/service/BookmarkServiceImpl.java
@@ -5,12 +5,15 @@ import org.omoknoone.onionhotsayyo.bookmark.aggregate.Bookmark;
 import org.omoknoone.onionhotsayyo.bookmark.dto.BookmarkDTO;
 import org.omoknoone.onionhotsayyo.bookmark.repository.BookmarkRepository;
 import org.omoknoone.onionhotsayyo.exceptions.BookmarkNotFoundException;
+import org.omoknoone.onionhotsayyo.follow.aggregate.Follow;
+import org.omoknoone.onionhotsayyo.follow.dto.FollowDTO;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.awt.print.Book;
 import java.util.List;
 
 @Service
@@ -29,16 +32,24 @@ public class BookmarkServiceImpl implements BookmarkService{
 	@Transactional
 	@Override
 	public void addBookmark(BookmarkDTO bookmarkDTO) {
-		log.info("북마크 추가 시작, 회원 ID {}");
+		if(!existsBookmark(bookmarkDTO)) {
+		log.info("북마크 추가 시작, 회원 ID {}", bookmarkDTO.getMemberId());
 		Bookmark bookmark = modelMapper.map(bookmarkDTO, Bookmark.class);
 		bookmarkRepository.save(bookmark);
+		} else {
+			throw new RuntimeException("이미 북마크 한 게시글입니다.");
+		}
 	}
 
 	@Transactional
 	@Override
 	public void removeBookmark(BookmarkDTO bookmarkDTO) {
+		if(existsBookmark(bookmarkDTO)){
 		Bookmark bookmark = modelMapper.map(bookmarkDTO, Bookmark.class);
 		bookmarkRepository.delete(bookmark);
+		} else {
+			throw new RuntimeException("삭제할 북마크가 없습니다.");
+		}
 	}
 
 	@Transactional(readOnly = true)
@@ -54,5 +65,19 @@ public class BookmarkServiceImpl implements BookmarkService{
 		log.info("회원 ID {}에 대한 북마크 목록 조회 완료: {}건 발견", memberId, bookmarks.size());
 
 		return bookmarks;
+	}
+
+
+	@Override
+	public boolean existsBookmark(BookmarkDTO bookmarkDTO) {
+		List<Bookmark> bookmarkList = bookmarkRepository.findAll();
+
+		for (Bookmark bookmark : bookmarkList) {
+			if(bookmark.getMemberId().equals(bookmarkDTO.getMemberId()) &&
+				bookmark.getPostId().equals(bookmarkDTO.getPostId())) {
+				return true;
+			}
+		}
+		return false;
 	}
 }
